Adds plotting functionality for JG experiments
Introduces a new plotting script and modifies existing scripts to generate plots for paper_Jacobi_Green experiments, comparing different preconditioners and resolutions.

The changes include:

- Creation of a new plotting script `exp_paper_JG_2D_elasticity_TO_1024_plot.py` that generates figures comparing the performance of different preconditioners.
- Modification of `exp_paper_JG_2D_elasticity_TO_1024.py` to support larger number of pixels and different experiment configurations.
- Update to the shell script `exp_paper_JG_elasticity_TO_1024.sh` to reduce the number of runs during the experiments
- Updates in the nonlinear elasticity examples

The aim of this modification is to improve the visualization of the experimental results and provide a more comprehensive analysis of the performance of different preconditioners.