Updates plotting and homogenization examples.
Modifies the homogenization examples to correctly display the phase field.

The phase field is now correctly passed as a field object to the plotting functions, ensuring that the plots are generated using the correct values.

Additionally, some minor adjustments were made to plotting parameters, such as letter offsets, to enhance visual clarity.

Some parameters are added/modified for JG preconditioner.